requirements.txt
flask
flask-cors
requests
pyravendb
openai
python-frontmatter